External commands: view and convert
The ExternalCommands class is used for the initiation of new external
commands like viewing PDF or DVI, convert documents from DVI to PDF or
PS, etc.

ExternalCommands create a new LogStore via the LogZone. Some checks are
made. If it's ok, then a new ExecuteCommand object is created, which
execute the command and write some messages on the LogStore.

It is possible to execute several commands at the same time (which was
not possible in LaTeXila 0.2).

For viewing and converting a document, the output of the command is not
showed because it's not interesting. The exit code is printed.