External commands: stop execution
Now the LogStore can have a reference to an ExternalCommand. This
reference is weak because ExternalCommand has also a reference to the
LogStore.
This reference is used for stopping the execution, since the only way to
access the ExternalCommand is via the LogStore.

LogStore has a "can_stop" property, so the log zone change the
sensitivity of the stop execution action when this property change (if
it is the current log store).
When the print_ouptut_exit () method is called, can_stop is set to
false.