Build Tools: load settings from XML file
Later it will be possible to modify the build tools settings with a GUI,
like in gedit-latex-plugin.

AppSettings is now a singleton, and the build tools settings are stored
there.
There are two placeholders in ui.xml, one in the Build menu, the other
in the main toolbar. So the build tools actions can be added or removed
dynamically.