Build tools: execute external commands
build_tools.vala contains several classes:
- BuildToolProcess (abstract)
- BuildToolRunner: subclass of BuildToolProcess
- PostProcessor (interface)
- GenericPostProcessor

A post processor is called only when the child process has finished,
when the exit code is known.

When a build tool is running, the GUI is not frozen (in LaTeXila 0.2
it was the case...) because the channels are read every 250 milliseconds
with a timeout.