Build tools: various improvements
When we click on a issue message that contains a filename, we jump to
this file. And if the is a line specified, go to this line.

Show/Hide bottom panel. When a build tool is executed the bottom panel
is automatically showed.

In the build view, when we click on a row that contains children,
expand or collapse this row.