Bug fix: critical messages with several windows
When we open a new window and close it, there were a lot of critical
messages when selecting text or click in the menu, ... in the first
window.

The problem was that a glade file was used for the mini file browser
toolbar (go home, go to parent directory, ...). So this glade file was
loaded several times, and its content displayed several times.

So the solution is to get rid of this glade file.