LaTeX menu: new items + some changes
Now the description of an item menu is displayed in the statusbar, so
it's better to indicate in the menu solely the latex command.
And the latex command must also be present in the description because if
this item is included in a toolbar, the user see only the description.