LaTeX menu: some changes
The indentation style is taken into account for list environments, which
makes an indentation for "\item".

Also, for characters sizes and font styles, the text inserted is
diffirent if there is no selection, or a selection on a single line, or
a selection on several lines. See insert_character_style ().