Set main file
If a document is splitted into several files, we can specify the main
file so when we run a build tool, it's executed on the main file.
Clean-up build files and view log take the main file into account too.

A comment like this is inserted in the file:

% mainfile: my_mainfile.tex

This comment must be present in the first or last three lines of the
file.