Completion: latex command names
We can type Ctrl+space or the beginning of a command like "\do" and the
auto-completion works. For interactive activation, 3 characters minimum
is required ('\' included). But it will be configurable.

The completion is only for latex command _names_, not for parameters.
The *.xml file contains data about possible values of parameters, so I
hope it will be soon implemented.

Also, the proposals are sorted in alphabetical order, but it's not
sorted by packages. It would be better to have one provider for each
latex package.