Some symbols regenerated
The 8 last operators (integrals and limits) are now smaller, so there
are more columns in the icon view.
The 16 first misc text symbols have now a bigger height, so there is
more space between the rows (before it was difficult to distinguish
some symbols, the height of the images was too small).