Completion tooltip: current arg in bold
Now the choices proposals are generated when needed. All the commands
that have one argument minimum are stored in an HashMap, the keys are
the command names.

Better detection of the argument number and type. For example:
\newtheorem{}[]{}[]
If we omit the first optional argument and we write:
\newtheorem{}{}[here]
Before, "here" was considered as the first optional argument, but in
reality it's the second.