Projects: the basis
A project is simply one directory with one main file. The other files in
this directory are belonging to the project.

The projects are saved in a XML file (in
~/.local/share/latexila/projects.xml). This file is parsed, and a
Gee.LinkedList contains all the projects, in AppSettings.

There is a new Document property: project-id.

The document tab label tooltip contains information about the main file.

Now I must implement the menu operations (new project, configure current
project, manage projects).