C code generated
Now it's possible to install latexila without the need to install Vala.

Since latexila requires most of the time a specific version of Vala (now
it's the 0.10.x release), it's easier to distribute the C code.

To regenerate the C code, use "-DBUILD_VALA=ON" in the cmake command.

The CMake Vala macros has been updated.