latex post processor
The purpose is to use directly "latex" or "pdflatex" commands.
The output is filtered to show only warnings, errors, badboxes and other
useful information. This filter is almost the same as in Kile, and was
already used in latexila 0.2.

Now the commands output is converted to utf-8 if needed (in
build_tool_runner). This conversion is made line by line, because if it
is done to the all output at once, there are some encodings troubles
with "latex" and "pdflatex" commands (with accents in the filename for
instance).