Build view: use weight property instead of markup
Before, the 'markup' was used for the text renderer. But there were some
problems with the 'all-output' post processor.

One solution would have been to escape the messages and still use
'markup', but it could be slow.

Since the markup is only used for putting a message in bold, we use the
'text' property instead of 'markup', and use the 'weight' property.