Build tools: better split of command arguments
To execute the command, we give an array of command arguments. But the
build tool contains only one string with the all command. So we have to
split it into arguments.

Before, the split was made only with spaces as delimiter. But some
arguments begin and end with ' or ". That's the case for example with
latexmk (here there are 4 arguments):

$ latexmk -e '$latex=q/latex %O -interaction=nonstopmode %S/' $filename