Latexmk post processor: option to show all info
With Latexmk we can have more information (which commands are executed
and the ouptut of each command), but most of the time the user wants to
see only the latex or pdflatex output.

So if the option is disabled, all the information will be displayed only
if an error occured during a command other than 'latex' or 'pdflatex'.