Structure: analyze a document (the basis)
There is a refresh button. When we click on it the current document is
analyzed and the structure is filled.

The items that are detected:
- simple commands like \name{text}
- TODO and FIXME comments

Tables and figures are NOT yet detected because it's a bit more complex
(find the caption etc).

There is a tooltip so we know what is the type of the item (there
is also the icon, but with the tooltip it's clearer).

The items are not sorted, it's a simple list without any hierarchy. The
sort will be made later in another commit.