Update the structure on tab changed
Now when several documents are opened at the same time, we jump to the
first doc, but we don't jump to the others. So the structure parsing is
done only on the first document in this case.

Also, an old bug is fixed: the swith-page signal was called two times
instead of one. Consequently, the structure update was called two times too.

And there are also some little code clean-up.