Unsaved document number: more robust
There was a strange bug since the rename of the property
'unsaved_document_n' to 'unsaved_document_num': the value was 0. With
some code inspection, the fact that it worked before was just pure
coincidence.

So now a lazy initialization is used, which is more robust (plus some
precaution about infinite loop, which is always better).