Structure: run the parsing in an IdleSource
So the UI is not frozen when the parsing is running.

When the function of the IdleSource is run, it runs until it's finished.
So now the parsing is splitted into chunks of 500 lines maximum. If the
user click on a menu, the menu will be displayed only when the 500 lines
are parsed.  But parsing 500 lines doesn't take a lot of time, so the
user see the menu almost directly.

If 500 lines is too big (e.g. for old computers), I can obviously reduce
the number. I first set it to 1000 and I didn't have any problem with my
computer (processor: AMD Turion(tm) 64 X2 Mobile Technology TL-62).