StructureModel: a custom Gtk.TreeModel
Using a separate tree (with GNode's) and then populate this tree into a
Gtk.TreeStore was not very clean, and the data were stored two times.

A custom TreeModel is a lot better, a bit more complicated, but for the
update on the fly of the structure (when the document is modified) it
will be really useful, mainly for performance reasons.