Structure: line-by-line parsing
It's between 5 and 6 times faster!

The TextMark's locations are now always at the beginning of lines. If
there are several items in the same line, they will point to the same
location. But it doesn't matter, what is important is to have the good
line.

It was not possible to have the exact location, because positions
returned by a regex is in byte, while creating a TextIter requires to
know the "UTF-8" position.