Structure: don't take into account commented markups
At first I wanted to display commented stuff in italic for example, so
we can easily uncomment the item.

But uncommenting the item can be difficult. If an entire section is
commented, we have to uncomment the entire section. But if it's just the
\section command which is commented, we must not uncomment the entire
section.

So the best is to use Edit -> Uncomment.