Fix structure update when active document changes
At first, I tried

    StructType? _current_list_type;

and set this variable to null if the simple list is hidden.
But when comparing this variable with another StructType, the comparison
was a "pointer comparison" (not the values). So there is a second boolean
variable:

    bool _list_is_hidden;

And it is better like that because a variable should be used only for
one purpose.