Structure: select the corresponding item in tree or list
If we select an item in the list, the corresponding item in the tree is
selected, and inversely.

To avoid infinite loop, a variable is used: _first_select.
Also, before selecting the corresponding item, we unselect all the rows.
Without that, the row we want to select is selected two times. For
example if we want to select row A but row B is already selected:

select row A
unselect row B
select row A

It's a bit stupid, and with the use of _first_select, it's buggy. So,
first, we unselect all the rows, so it becomes:

unselect row B
select row A

And everything is OK.