Cleanup Build Files: take into account projects
If a project is defined, there were some remaining *.aux files.

So in this case, instead of replacing the extension of the main file
to see if the file exist and remove it, search directly all files
which have the extension (in all sub-directories), and remove them.

The code has been moved to a new class, so it's cleaner.