Build View: text is selectable, so we can copy/paste it
My first idea was to add a menu on right click, to copy the row contents
in the clipboard. But it's easier to set the text editable. If the user
edit the text, the changes are not taken into account. And it's better
like this because for copy/paste a command, we can select the command
without the dollar sign.