Structure: refactoring: more re-usable parsing functions
The parsing functions can be used to re-parse a specific line without
the aim of parsing all the document and filling the structure model.

It will be useful for some structure actions (cut, copy, select, ...),
because for these actions, we must know the exact places where the item
begins and where it ends. Currently, we only know the line number,
because the TextMark is always placed at the beginning of the line.