Various UI improvements
Build Tools list: the columns now have a title: active, and label.
Double-click on the active column (with the checkbutton) won't open the
properties.

The Structure have tree lines and there are two vertical separators in
the mini toolbar.

And the editor font button in the preferences dialog doesn't expand (the
window becomes quite large, so it was a bit ugly).