Use ITS Tool for the templates
The templates are encoded in XML. Each template is composed of several
chunks.

For all the templates except the letter, there are three chunks, but
only the second chunk is translatable (for the babel package or
equivalent).

For the letter, there is one big chunk with all the template. Decompose
the letter into several strings to translate would not have been enough
flexible, because in some languages we want to use for example another
document class, with other commands, etc.

There is a script: gen_tex.sh. This script removes the XML markups and
generate .tex files.

What is installed is only the .tex files. So there is one more step to
do here (call the script).

C/*.xml -> C/*.tex
lang.po -> lang.mo -> lang-xml/*.xml -> lang/*.tex