Structure: optimize insertion of figures and tables
As soon as we find the \begin{env}, we insert the item in the model.
When we find the first caption, we store it in a variable, and when we
find the corresponding \end{env}, we modify the data of the item.

This is when the data is modified that the items between the \begin and
the \end are moved as children of the environment.

This is a lot faster than waiting the \end{env} and insert it in middle.