Add the *.pot files in po/POTFILES.in
For the build tools and the templates, ITS Tool is used with custom .its
files. The solution to integrate these translations with Damned Lies is
to simply add the *.pot files in the POTFILES.in file. The drawback is
to keep the *.pot files in Git, and update them when there is a
modification.