CMake option: WITH_GNOME, for gsettings-desktop-schemas
Before, the dependency to gsettings-desktop-schemas was not explicit.
During the cmake command, if g-d-s was already installed, then we used
it, if not, we didn't depend on it.

But if the user uninstall g-d-s afterwards, latexila is broken.

Now, the dependency is explicit. And by default, it is disabled. If you
want to enable it, run cmake with:

-DWITH_GNOME=ON