Completion: simplify match()
Because show_no_proposals() (which is called in populate()) has the same
effect as returning false in match().

And now it's even more logic for the minimum number of characters needed
to trigger interactive completion: before, when there were proposals
thanks to interactive completion, when we deleted characters, proposals
were still there even when the number of chars is below the minimum.

So now to have completion below the minimum nb of chars set for
interactive completion, we have to request it explicitly (Ctrl+space).