Simplify search and replace
Altough the search and replace is more convenient than in Gedit,
the code is quite ugly and is very slow for big documents. It will be
better in the long term to make the gedit code more reusable, improve
it, and use it in latexila.

So for now the code has been simplified (the labels that show
information like the number of matches, whether the word is not found, …
are removed). It's simpler to remove this feature than to adapt the UI.