Structure: remove "expand all" from the toolbar
This action is a bit useless because by default all the items are
already expanded.

Another (better) reason to remove this action is that the mini-toolbar
had too many actions. On a small screen, we want to have the side panel
as small as possible, but with a long mini-toolbar is not possible in
GTK+ 3.