Symbols: refactoring and use GResource
The most used symbols are disabled. This will be the next part.

The "make install" is a bit shorter thanks to GResource ;) The 614
symbols are now inside the executable.

There is now the SymbolsView class, which contain only the UI to display
the symbols in the side panel.

There is also a private class NormalSymbols (child of ListStore) to get
the model of a "normal" store (i.e. not the most used symbols store).
The MostUsedSymbols class will be adapted in the same way.

All the data about the symbols are in XML files (loaded with GResource).