Most used symbols: refactoring
The ListStore is sorted in descending order (with the column containing
the number of times the symbol is used). And then a TreeModelFilter is
used to take only the N first symbols, depending on the preferences.

In the XML, only two things are saved now for each symbol: the id, and
the number of times the symbol has been used. Before, the latex command
and the package were also saved, but it will be cleaner to take these
information from the other categories (this is not already done). If for
example the information is modified (e.g. because the latex command is
not correct, or the package required was missing, ...), it will get the
right information instead of keeping the old data in the XML.