Create the build command runner
Initially, the build command runner (it was maybe another name, I don't
recall exactly) was the parent class of the build tool runner. Then the
two classes has been merged. And now instead of using inheritance,
delegation is used.

It is a first step to clean and refactor the build tool runner.