Create the BuildJobRunner class
To summarize a little, there is the BottomPanel containing the
BuildView. When a build tool is executed, a BuildToolRunner is created,
which fill the build view. A BuildToolRunner creates, for each build
job, a BuildJobRunner. In turn, a BuildJobRunner creates a
BuildCommandRunner, or (in the future), call gtk_show_uri(), and creates
a PostProcessor with the output of the command.