Build tool runner: remove Gtk.Action parameter
I think the purpose was to set the Gtk.Action as insensitive directly,
without passing by a signal (here the 'finished' signal), to be more
efficient. There was probably a problem: when we click to abort the
build tool, the button was not set as insensitive directly, so we could
click several times on the abort button. Now the problem seems to be
gone.