Build tool dialog: the UI for opening files
Currently, the "$view" placeholder is used to open files. xdg-open is
used for that, but the plan is to use gtk_show_uri(). It would have been
possible to keep the $view placeholder, analyze the commands and execute
gtk_show_uri() when we see a $view. But the problem is that the Post
Processor is not used when gtk_show_uri() is called. A solution would be
to make the post processor insensitive when $view is present in the
command. But I don't like this.

With this commit, it is IMHO easier: first, the build jobs (the
commands) are executed. And if all the commands succeed, then the
specified files are opened.

It is not useful to open a file before executing a command. But if the
user wants absolutely do that, it is still possible to add a command
with xdg-open or another application's name.

So there is clearly a distinction between executing a command, and
opening a file.