Build Tools: show/hide detailed messages
There is a problem with the (owned) Node<BuildMsg?>. The same problem
when using the GLib.List. It is better to use Gee in Vala, but Gee
doesn't have a Node interface. So I'll use Gee.List, and in the
structure have another Gee.List with the children, to form the tree.