File browser: automatic refresh
By using a FileMonitor. There is a one-second delay, so if there are
many changes at once, the function is called only one time.

This is because the function called clean completely the TreeStores and
re-enumerate all the files in the directory. It is possible to update
incrementally the TreeStores, but it is more complicated, and it is not
a priority.