Simplify the file browser settings
Before, it was possible to list the file's extensions that are displayed
in the file browser. But is it really useful?

What most users want, I think, is simply to be able to hide the build
files (those for clean-up: .log .toc …). By default the build files are
hidden.

The setting for showing/hiding hidden files is kept (the files that
begins with a dot).