Fix inline toolbars style: use a Box instead of a Grid
With a GtkGrid the inline toolbars style is not perfect. The line
separating the TreeView and the Toolbar is thicker. And the corners at
the bottom of the toolbars are not rounded.

With a vertical GtkBox, it's OK.

The bug should be reported if it is not already the case.