Autotools: handle the templates
Instead of generating .tex templates for each language with ITS Tool,
the XML files are directly read in Vala, and the chunks that must be
translated, are translated there.

The Vala code is more complicated, but the Autotools/CMake "code" is
easier. Translating the CMake macro for the Autotools (a M4 macro for
example) would have been more complicated I think (at least for me).