Fix show/hide of main window components at startup
The visibility of the main and edit toolbars, the side panel and the
bottom panel was not always set correctly at startup.

Now when a component is by default shown or hide, the corresponding
value is set to the GtkAction, so when this value is modified later by
the show_or_hide_widgets() function, if the value is not the same, the
notify signal is correctly emitted.