Better way for shrinking the window when unmaximized
Before, if the window was initially maximized, the default size of the
window was set smaller (100 pixels less) than the maximized size, with
the set_default_size() function.

But due to this smaller size, the paned positions were not restored
correctly. When restoring the positions, the window is not already
shown, and so the default size is taken. When the window is shown and
then maximized, the paned position has 100 more pixels. After several
restarts, the bottom panel was completely hidden.