Change the application icons
This was the first icons made by Éric Forgeot. There is an SVG version.
For big sizes (e.g. in gnome-shell) it is really nice, but for small
sizes (e.g. in Xfce), it is a bit ugly. But the previous icon (made by
Ann Melnichuk) was also a bit ugly for small sizes.