Backward search: check that evince is the default document viewer
If evince is not the default document viewer for PDF files, the
backward search with synctex is now disabled.

There was a bug with the build tools: when opening a PDF file, the
document was always opened by evince. And if the default document viewer
was not evince, the document was also opened with the other document
viewer. So the PDF file was opened twice.

Thanks to Richard Garavuso for the bug report.