Use upstream GtkSpell, version >= 3.0.3 required
When developing LaTeXila 2.6 (first version with GTK+ 3), GtkSpell was
not well maintained, there was no official version for GTK+ 3, GtkSpell
was not GObjectified, etc. That's why there was a copy of the GtkSpell
source code in latexila.

Now GtkSpell is well maintained, it has been GObjectified, and several
versions for GTK+ 3 have been released, including new features.