Search: improve replace behavior
If we click on the replace button, if no occurrence is selected, go to
the next occurrence. If an occurrence is selected, just replace it, do
not move to the next occurrence. So when we have replaced an occurrence,
we have the time to see if the replacement is correct, and fix it if we
are not happy with the result. I prefer this behavior.