Search: print match infos
Print the current match position, and the total number of matches.
In gedit these numbers are integrated in the text entry, with a
GdTaggedEntry. But I'm lazy, I don't like doing two times the same
thing. So here the information is printed in a simple label, on the
right of the close button. The advantage with the label is that the
information is clearer: "Match 3 of 10", "10 matches" and "Not found".