Create liblatexila with empty GObject class
The liblatexila is not a real library like GTK+, it is just an internal
library. The code will be written in C in this directory. The Vala code
in the parent directory can use the liblatexila, but I'll try to not do
the reverse (use the Vala code in liblatexila) so the liblatexila is
self-contained and could be reused easily by other projects (a gedit
plugin for example).

Why the C language? Well, I now prefer the C language. The Vala syntax
is nice, but it adds another layer. Another layer of potential bugs.
Another layer of maintenance. Another layer for the API documentation.
The GNOME documentation with GTK-Doc is primarily written for the C
language. And the verbosity of the C language has advantages
("greppable" code, code more self-documented).